refactor: moved configuration mapping related classes to single location
Refs: XRDDEV-2534
