+++ src/common/common-globalconf/build.gradle
@@ -3,7 +3,11 @@ plugins {
 }
 
 dependencies {
+    annotationProcessor(libs.mapstructProcessor)
+    annotationProcessor(libs.lombokMapstructBinding)
+
     implementation(platform(libs.springBoot.bom))
+    implementation(libs.mapstruct)
 
     api project(":common:common-core")
     api project(":common:common-domain")
