+++ src/common/common-globalconf/src/test/java/ee/ria/xroad/common/conf/globalconf/PrivateParametersV3MarshallerTest.java
@@ -1,21 +1,21 @@
 /*
  * The MIT License
- * <p>
+ *
  * Copyright (c) 2019- Nordic Institute for Interoperability Solutions (NIIS)
  * Copyright (c) 2018 Estonian Information System Authority (RIA),
  * Nordic Institute for Interoperability Solutions (NIIS), Population Register Centre (VRK)
  * Copyright (c) 2015-2017 Estonian Information System Authority (RIA), Population Register Centre (VRK)
- * <p>
+ *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
  * in the Software without restriction, including without limitation the rights
  * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
  * copies of the Software, and to permit persons to whom the Software is
  * furnished to do so, subject to the following conditions:
- * <p>
+ *
  * The above copyright notice and this permission notice shall be included in
  * all copies or substantial portions of the Software.
- * <p>
+ *
  * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
  * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
  * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
@@ -24,10 +24,8 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
+package ee.ria.xroad.common.conf.globalconf;
 
-package org.niis.xroad.cs.admin.globalconf.generator;
-
-import ee.ria.xroad.common.conf.globalconf.PrivateParameters;
 import ee.ria.xroad.common.identifier.ClientId;
 
 import jakarta.xml.bind.MarshalException;
@@ -36,11 +34,10 @@ import org.junit.jupiter.api.Test;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 
-class PrivateParametersV2MarshallerTest {
-
+class PrivateParametersV3MarshallerTest {
     private final ClientId clientId = ClientId.Conf.create("CS", "CLASS", "CODE");
 
-    private final PrivateParametersV2Marshaller marshaller = new PrivateParametersV2Marshaller();
+    private final PrivateParametersV3Marshaller marshaller = new PrivateParametersV3Marshaller();
 
     @Test
     void marshall() {
@@ -87,5 +84,4 @@ class PrivateParametersV2MarshallerTest {
 
         assertThrows(MarshalException.class, () -> marshaller.marshall(privateParams));
     }
-
 }
