+++ src/common/common-globalconf/src/test/java/ee/ria/xroad/common/conf/globalconf/SharedParametersV2ToXmlConverterTest.java
@@ -24,11 +24,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.niis.xroad.cs.admin.globalconf.generator;
+package ee.ria.xroad.common.conf.globalconf;
 
-import ee.ria.xroad.common.conf.globalconf.SharedParameters;
-import ee.ria.xroad.common.conf.globalconf.sharedparameters.v3.ObjectFactory;
-import ee.ria.xroad.common.conf.globalconf.sharedparameters.v3.SharedParametersTypeV3;
+import ee.ria.xroad.common.conf.globalconf.sharedparameters.v2.ObjectFactory;
+import ee.ria.xroad.common.conf.globalconf.sharedparameters.v2.SharedParametersTypeV2;
 import ee.ria.xroad.common.identifier.ClientId;
 import ee.ria.xroad.common.util.CryptoUtils;
 
@@ -55,13 +54,10 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
 @Slf4j
-class SharedParametersV3ConverterTest {
+class SharedParametersV2ToXmlConverterTest {
 
     private static final Map<String, String> FIELD_NAME_MAP = Map.ofEntries(
             entry("securityServer", "securityServers"),
-            entry("source", "sources"),
-            entry("internalVerificationCert", "internalVerificationCerts"),
-            entry("externalVerificationCert", "externalVerificationCerts"),
             entry("approvedCA", "approvedCAs"),
             entry("approvedTSA", "approvedTSAs"),
             entry("member", "members"),
@@ -77,7 +73,7 @@ class SharedParametersV3ConverterTest {
     @Test
     void shouldConvertAllFields() throws IOException, OperatorCreationException {
         var sharedParameters = getSharedParameters();
-        var xmlType = SharedParametersV3Converter.INSTANCE.convert(sharedParameters);
+        var xmlType = SharedParametersV2ToXmlConverter.INSTANCE.convert(sharedParameters);
 
         var conf = RecursiveComparisonConfiguration.builder()
                 .withIntrospectionStrategy(compareRenamedFields())
@@ -86,8 +82,7 @@ class SharedParametersV3ConverterTest {
                         "securityServers.authCerts",
                         "members.id",
                         "members.subsystems.id",
-                        "centralService",
-                        "any"
+                        "centralService"
                 )
                 .withEqualsForFields((a, b) -> new BigInteger(a.toString()).compareTo(new BigInteger(b.toString())) == 0,
                         "globalSettings.ocspFreshnessSeconds")
@@ -105,12 +100,12 @@ class SharedParametersV3ConverterTest {
 
         assertIdReferences(xmlType);
         assertThat(xmlType.getSecurityServer().get(0).getAuthCertHash().get(0))
-                .isEqualTo(CryptoUtils.certHash(sharedParameters.getSecurityServers().get(0).getAuthCertHashes().get(0)));
+                .isEqualTo(CryptoUtils.certSha1Hash(sharedParameters.getSecurityServers().get(0).getAuthCertHashes().get(0)));
     }
 
     @Test
     void shouldBeAbleToMarshall() throws JAXBException {
-        var xmlType = SharedParametersV3Converter.INSTANCE.convert(getSharedParameters());
+        var xmlType = SharedParametersV2ToXmlConverter.INSTANCE.convert(getSharedParameters());
 
         JAXBContext jaxbContext = JAXBContext.newInstance(ObjectFactory.class);
         var writer = new StringWriter();
@@ -125,10 +120,10 @@ class SharedParametersV3ConverterTest {
 
     @Test
     void shouldReturnNullWhenInputIsNull() {
-        assertThat(SharedParametersV3Converter.INSTANCE.convert((SharedParameters) null)).isNull();
+        assertThat(SharedParametersV2ToXmlConverter.INSTANCE.convert((SharedParameters) null)).isNull();
     }
 
-    private static void assertIdReferences(SharedParametersTypeV3 xmlType) {
+    private static void assertIdReferences(SharedParametersTypeV2 xmlType) {
         var ownerMember = xmlType.getMember().get(0);
         var client = ownerMember.getSubsystem().get(0);
 
@@ -154,19 +149,11 @@ class SharedParametersV3ConverterTest {
     }
 
     private static SharedParameters getSharedParameters() {
-        return new SharedParameters("INSTANCE", getConfigurationSources(), List.of(getApprovedCA()),
+        return new SharedParameters("INSTANCE", List.of(), List.of(getApprovedCA()),
                 List.of(new SharedParameters.ApprovedTSA("tsa-name", "tsa-url", "tsa cert".getBytes(UTF_8))),
                 getMembers(), List.of(getSecurityServer()), List.of(new SharedParameters.GlobalGroup("group-code",
                 "group-description", List.of(subsystemId(memberId(), "SUB1")))),
                 new SharedParameters.GlobalSettings(List.of(getMemberClass()), 333));
-    }
-
-    private static List<SharedParameters.ConfigurationSource> getConfigurationSources() {
-        var configurationSource = new SharedParameters.ConfigurationSource();
-        configurationSource.setAddress("cs");
-        configurationSource.setInternalVerificationCerts(List.of("internal-conf-singing-cert".getBytes(UTF_8)));
-        configurationSource.setExternalVerificationCerts(List.of("external-conf-singing-cert".getBytes(UTF_8)));
-        return List.of(configurationSource);
     }
 
     private static SharedParameters.ApprovedCA getApprovedCA() {
@@ -176,7 +163,6 @@ class SharedParametersV3ConverterTest {
         approvedCA.setTopCA(getCaInfo());
         approvedCA.setCertificateProfileInfo("certificateProfileInfo");
         approvedCA.setIntermediateCas(List.of(getCaInfo()));
-        approvedCA.setAcmeServer(new SharedParameters.AcmeServer("http://testca.com/acme", "192.99.88.7", "1", "2"));
         return approvedCA;
     }
 
@@ -219,10 +205,6 @@ class SharedParametersV3ConverterTest {
     }
 
     private static ClientId.Conf subsystemId(ClientId.Conf clientId, String subsystemCode) {
-        return ClientId.Conf.create(clientId.getXRoadInstance(),
-                clientId.getXRoadInstance(),
-                clientId.getMemberCode(),
-                subsystemCode);
+        return ClientId.Conf.create(clientId.getXRoadInstance(), clientId.getXRoadInstance(), clientId.getMemberCode(), subsystemCode);
     }
-
 }
