+++ src/common/common-globalconf/src/test/java/ee/ria/xroad/common/conf/globalconf/SharedParametersV3ToXmlConverterTest.java
@@ -24,11 +24,10 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.niis.xroad.cs.admin.globalconf.generator;
+package ee.ria.xroad.common.conf.globalconf;
 
-import ee.ria.xroad.common.conf.globalconf.SharedParameters;
-import ee.ria.xroad.common.conf.globalconf.sharedparameters.v4.ObjectFactory;
-import ee.ria.xroad.common.conf.globalconf.sharedparameters.v4.SharedParametersTypeV4;
+import ee.ria.xroad.common.conf.globalconf.sharedparameters.v3.ObjectFactory;
+import ee.ria.xroad.common.conf.globalconf.sharedparameters.v3.SharedParametersTypeV3;
 import ee.ria.xroad.common.identifier.ClientId;
 import ee.ria.xroad.common.util.CryptoUtils;
 
@@ -55,7 +54,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 
 @Slf4j
-class SharedParametersV4ConverterTest {
+class SharedParametersV3ToXmlConverterTest {
 
     private static final Map<String, String> FIELD_NAME_MAP = Map.ofEntries(
             entry("securityServer", "securityServers"),
@@ -77,7 +76,7 @@ class SharedParametersV4ConverterTest {
     @Test
     void shouldConvertAllFields() throws IOException, OperatorCreationException {
         var sharedParameters = getSharedParameters();
-        var xmlType = SharedParametersV4Converter.INSTANCE.convert(sharedParameters);
+        var xmlType = SharedParametersV3ToXmlConverter.INSTANCE.convert(sharedParameters);
 
         var conf = RecursiveComparisonConfiguration.builder()
                 .withIntrospectionStrategy(compareRenamedFields())
@@ -110,7 +109,7 @@ class SharedParametersV4ConverterTest {
 
     @Test
     void shouldBeAbleToMarshall() throws JAXBException {
-        var xmlType = SharedParametersV4Converter.INSTANCE.convert(getSharedParameters());
+        var xmlType = SharedParametersV3ToXmlConverter.INSTANCE.convert(getSharedParameters());
 
         JAXBContext jaxbContext = JAXBContext.newInstance(ObjectFactory.class);
         var writer = new StringWriter();
@@ -125,10 +124,10 @@ class SharedParametersV4ConverterTest {
 
     @Test
     void shouldReturnNullWhenInputIsNull() {
-        assertThat(SharedParametersV4Converter.INSTANCE.convert((SharedParameters) null)).isNull();
+        assertThat(SharedParametersV3ToXmlConverter.INSTANCE.convert((SharedParameters) null)).isNull();
     }
 
-    private static void assertIdReferences(SharedParametersTypeV4 xmlType) {
+    private static void assertIdReferences(SharedParametersTypeV3 xmlType) {
         var ownerMember = xmlType.getMember().get(0);
         var client = ownerMember.getSubsystem().get(0);
 
@@ -157,21 +156,8 @@ class SharedParametersV4ConverterTest {
         return new SharedParameters("INSTANCE", getConfigurationSources(), List.of(getApprovedCA()),
                 List.of(new SharedParameters.ApprovedTSA("tsa-name", "tsa-url", "tsa cert".getBytes(UTF_8))),
                 getMembers(), List.of(getSecurityServer()), List.of(new SharedParameters.GlobalGroup("group-code",
-                        "group-description", List.of(subsystemId(memberId(), "SUB1")))),
+                "group-description", List.of(subsystemId(memberId(), "SUB1")))),
                 new SharedParameters.GlobalSettings(List.of(getMemberClass()), 333));
-
-//        parameters.setInstanceIdentifier("INSTANCE");
-//        parameters.setSources(getConfigurationSources());
-//        parameters.setApprovedCAs(List.of(getApprovedCA()));
-//        parameters.setApprovedTSAs(List.of(new SharedParameters.ApprovedTSA("tsa-name",
-//                "tsa-url",
-//                "tsa cert".getBytes(UTF_8))));
-//        parameters.setMembers(getMembers());
-//        parameters.setSecurityServers(List.of(getSecurityServer()));
-//        parameters.setGlobalGroups(List.of(new SharedParameters.GlobalGroup("group-code", "group-description",
-//                List.of(subsystemId(memberId(), "SUB1")))));
-//        parameters.setGlobalSettings(new SharedParameters.GlobalSettings(List.of(getMemberClass()), 333));
-//        return parameters;
     }
 
     private static List<SharedParameters.ConfigurationSource> getConfigurationSources() {
